refactor(evm): optimize condition serialization with linear-pass topology
Refactors the condition tree processing pipeline to use a single O(N) backward pass for topology analysis, replacing redundant recursive traversals.

+++ packages/evm/contracts/types/Condition.sol
@@ -29,7 +29,7 @@ struct Condition {
 // used for ABI encoding a scope config tree
 // (ABI does not support recursive types)
 struct ConditionFlat {
-    uint8 parent;
+    uint16 parent;
     Encoding paramType;
     Operator operator;
     bytes compValue;
